+++ navground_sim_py/navground/sim/state_estimations/__init__.py
@@ -1,4 +1,5 @@
 from .._navground_sim import (BoundarySensor, BoundedStateEstimation,
                               SensorCombination, DiscsStateEstimation,
-                              LidarStateEstimation, Sensor)
+                              LidarStateEstimation, Sensor,
+                              OdometryStateEstimation)
 from .lidar import PyLidarStateEstimation